Fix auth URL
The way that the auth url was being used by AuthorizationService was too restrictive
and made too many assumptions about the format of the URL; GitLab, for example, uses
a different URL format and does not name the service the same as DockerHub does.

These changes add a parameter, 'auth_service_name', which allows custom service names
that are not simply the same as the registry's host name (e.g. GitlLab uses
'container_registry'), and also changes the URL so that the 'v2/token' location is
not necessarily assumed.

*This is necessarily a breaking change for existing consumers and will require a majorversion bump*.